Split tabs into seperate files
