+++ src/openloco/company.cpp
@@ -76,4 +76,11 @@ namespace openloco
         args.push(performanceIndex);
         args.push(getCorporateRatingAsStringId(performanceToRating(performanceIndex)));
     }
+
+    bool company::isVehicleIndexUnlocked(uint8_t vehicleIndex)
+    {
+        auto vehicleTypeIndex = vehicleIndex >> 5;
+
+        return (unlocked_vehicles[vehicleTypeIndex] & (1 << (vehicleIndex & 0x1F))) != 0;
+    }
 }
