+++ src/openloco/openloco.vcxproj
@@ -92,7 +92,11 @@
     <ClCompile Include="windows\CompanyFaceSelection.cpp" />
     <ClCompile Include="windows\CompanyList.cpp" />
     <ClCompile Include="windows\CompanyWindow.cpp" />
-    <ClCompile Include="windows\constructionwnd.cpp" />
+    <ClCompile Include="windows\construction\Common.cpp" />
+    <ClCompile Include="windows\construction\ConstructionTab.cpp" />
+    <ClCompile Include="windows\construction\OverheadTab.cpp" />
+    <ClCompile Include="windows\construction\SignalTab.cpp" />
+    <ClCompile Include="windows\construction\StationTab.cpp" />
     <ClCompile Include="windows\EditKeyboardShortcut.cpp" />
     <ClCompile Include="windows\Error.cpp" />
     <ClCompile Include="windows\IndustryWindow.cpp" />
@@ -234,6 +238,7 @@
     <ClInclude Include="widget.h" />
     <ClInclude Include="win32.h" />
     <ClInclude Include="window.h" />
+    <ClInclude Include="windows\construction\Construction.h" />
     <ClInclude Include="windows\toolbar_top_common.h" />
   </ItemGroup>
   <ItemGroup>
